This function reads Arduino serial output and outputs an nx2 array with the name of your choosing.
To run this file, ensure that the Arduino is flashed with the EncoderMillis code, with the serial monitor closed.

Note that this code is HARD CODED to the serial port on the left side of my Macbook - if you want to use this, you need to change the 
serial1 = serial('/dev/tty.usbmodem1411');
line to reflect the serial port on your computer (think COM1 for Windows, an alternate serial port for Macs, or something else entirely for Linux)

To run this code after altering the serial input line, type:
'abc' = read_encoder();
where 'abc' is the name of the array you'd like.